# backend/src/models/task.py
"""Task model for managing user tasks."""
from sqlalchemy import Column, Integer, String, Text, Boolean, DateTime, func
from src.database import Base


class Task(Base):
    """Task model for focus sessions."""
    __tablename__ = "tasks"

    id = Column(Integer, primary_key=True, index=True)
    title = Column(String(255), nullable=False)
    description = Column(Text, nullable=True)
    completed = Column(Boolean, default=False, nullable=False)
    order = Column(Integer, default=0, nullable=False)
    created_at = Column(DateTime(timezone=True), server_default=func.now(), nullable=False)
    updated_at = Column(DateTime(timezone=True), onupdate=func.now(), nullable=True)