# specs/001-focus-flow/contracts/openapi.yaml
openapi: 3.0.3
info:
  title: FocusFlow API
  description: |
    ADHD-friendly focus tool API providing task management, Pomodoro timer sessions,
    habit tracking, and brain dump functionality.

    ## Key Features
    - **Task Management**: CRUD operations for tasks and micro-steps with drag-and-drop reordering
    - **Pomodoro Sessions**: Focus and break timer management
    - **Brain Dump**: Quick capture of intrusive thoughts
    - **Habit Tracking**: Daily habits with streak calculation
    - **User Settings**: Customizable focus/break durations and themes

    ## Design Principles
    - RESTful API design with standard HTTP methods
    - Idempotent operations for offline sync reliability
    - Comprehensive error responses
    - Optimistic locking support (ETags)

  version: 1.0.0
  contact:
    name: FocusFlow API Support
    email: support@focusflow.app
servers:
  - url: http://localhost:8000/api/v1
    description: Local development server
  - url: http://localhost:8000/api/v1
    description: Production server (placeholder)
...
paths:
...
  /tasks:
    get:
      summary: List all tasks
      description: Retrieve all tasks with optional filtering and sorting. Includes nested micro-steps.
      tags: [tasks]
      parameters:
        - name: completed
          in: query
          description: Filter by completion status
          schema:
            type: boolean
        - name: sort_by
          in: query
          description: Sort field
          schema:
            type: string
            enum: [created_at, updated_at, order, title]
            default: order
        - name: sort_order
          in: query
          description: Sort direction
          schema:
            type: string
            enum: [asc, desc]
            default: asc
      responses:
        '200':
          description: Successful response
          content:
            application/json:
              schema:
                type: object
                properties:
                  tasks:
                    type: array
                    items:
                      $ref: '#/components/schemas/Task'
                  total:
                    type: integer
                    description: Total number of tasks
              example:
                tasks:
                  - id: "task-123"
                    title: "Write research report"
                    description: null
                    completed: false
                    order: 0
                    created_at: "2026-01-06T12:00:00Z"
                    updated_at: "2026-01-06T12:00:00Z"
                    micro_steps: []
                total: 1

    post:
      summary: Create a new task
      description: Create a task with optional micro-steps
      tags: [tasks]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/TaskCreate'
            example:
              title: "Complete project proposal"
              description: "Draft and review proposal for Q2 project"
      responses:
        '201':
          description: Task created successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Task'
        '400':
          $ref: '#/components/responses/BadRequest'
        '422':
          $ref: '#/components/responses/ValidationError'
  /tasks/{task_id}:
    get:
      summary: Get task by ID
      tags: [tasks]
      parameters:
        - $ref: '#/components/parameters/TaskId'
      responses:
        '200':
          description: Successful response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Task'
        '404':
          $ref: '#/components/responses/NotFound'

    put:
      summary: Update task
      description: Update task fields. Supports partial updates.
      tags: [tasks]
      parameters:
        - $ref: '#/components/parameters/TaskId'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/TaskUpdate'
            example:
              title: "Complete project proposal (revised)"
              completed: false
      responses:
        '200':
          description: Task updated successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Task'
        '404':
          $ref: '#/components/responses/NotFound'
        '422':
          $ref: '#/components/responses/ValidationError'

    delete:
      summary: Delete task
      description: Delete task and all associated micro-steps (cascade)
      tags: [tasks]
      parameters:
        - $ref: '#/components/parameters/TaskId'
      responses:
        '204':
          description: Task deleted successfully
        '404':
          $ref: '#/components/responses/NotFound'
...
  /pomodoro/sessions:
    get:
      summary: List Pomodoro sessions
      description: Retrieve session history with optional filtering
      tags: [pomodoro]
      parameters:
        - name: task_id
          in: query
          description: Filter by task ID
          schema:
            type: string
            format: uuid
        - name: session_type
          in: query
          description: Filter by session type
          schema:
            type: string
            enum: [focus, break]
        - name: state
          in: query
          description: Filter by state
          schema:
            type: string
            enum: [running, paused, completed, cancelled]
        - name: limit
          in: query
          description: Number of sessions to return
          schema:
            type: integer
            default: 50
            maximum: 100
      responses:
        '200':
          description: Successful response
          content:
            application/json:
              schema:
                type: object
                properties:
                  sessions:
                    type: array
                    items:
                      $ref: '#/components/schemas/PomodoroSession'
                  total:
                    type: integer

    post:
      summary: Start a new Pomodoro session
      tags: [pomodoro]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/PomodoroSessionCreate'
            example:
              task_id: "task-123"
              session_type: "focus"
              duration_minutes: 25
      responses:
        '201':
          description: Session started
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/PomodoroSession'
        '400':
          $ref: '#/components/responses/BadRequest'
  /pomodoro/sessions/{session_id}:
    get:
      summary: Get session by ID
      tags: [pomodoro]
      parameters:
        - name: session_id
          in: path
          required: true
          schema:
            type: string
            format: uuid
      responses:
        '200':
          description: Successful response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/PomodoroSession'
        '404':
          $ref: '#/components/responses/NotFound'

    put:
      summary: Update session state
      description: Pause, resume, complete, or cancel a session
      tags: [pomodoro]
      parameters:
        - name: session_id
          in: path
          required: true
          schema:
            type: string
            format: uuid
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                state:
                  type: string
                  enum: [paused, running, completed, cancelled]
              required: [state]
            example:
              state: "paused"
      responses:
        '200':
          description: Session updated
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/PomodoroSession'
        '400':
          description: Invalid state transition
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '404':
          $ref: '#/components/responses/NotFound'
...
components:
  parameters:
    TaskId:
      name: task_id
      in: path
      required: true
      description: Task UUID
      schema:
        type: string
        format: uuid
  schemas:
    Task:
      type: object
      properties:
        id:
          type: string
          format: uuid
        title:
          type: string
          maxLength: 500
        description:
          type: string
          nullable: true
        completed:
          type: boolean
        order:
          type: integer
          minimum: 0
        created_at:
          type: string
          format: date-time
        updated_at:
          type: string
          format: date-time
        micro_steps:
          type: array
          items:
            $ref: '#/components/schemas/MicroStep'
      required:
        - id
        - title
        - completed
        - order
        - created_at
        - updated_at
    TaskCreate:
      type: object
      properties:
        title:
          type: string
          minLength: 1
          maxLength: 500
        description:
          type: string
          maxLength: 10000
          nullable: true
      required:
        - title
    TaskUpdate:
      type: object
      properties:
        title:
          type: string
          minLength: 1
          maxLength: 500
        description:
          type: string
          maxLength: 10000
          nullable: true
        completed:
          type: boolean
        order:
          type: integer
          minimum: 0
    MicroStep:
      type: object
      properties:
        id:
          type: string
          format: uuid
        task_id:
          type: string
          format: uuid
        title:
          type: string
          maxLength: 500
        completed:
          type: boolean
        order:
          type: integer
          minimum: 0
        created_at:
          type: string
          format: date-time
...
    PomodoroSession:
      type: object
      properties:
        id:
          type: string
          format: uuid
        task_id:
          type: string
          format: uuid
          nullable: true
        session_type:
          type: string
          enum: [focus, break]
        duration_minutes:
          type: integer
          minimum: 1
        state:
          type: string
          enum: [running, paused, completed, cancelled]
        started_at:
          type: string
          format: date-time
        completed_at:
          type: string
          format: date-time
          nullable: true
        paused_duration_ms:
          type: integer
          description: Total time paused in milliseconds
    PomodoroSessionCreate:
      type: object
      properties:
        task_id:
          type: string
          format: uuid
          nullable: true
        session_type:
          type: string
          enum: [focus, break]
        duration_minutes:
          type: integer
          minimum: 1
          maximum: 60
      required:
        - session_type
        - duration_minutes
...
    Error:
      type: object
      properties:
        error:
          type: string
        detail:
          type: string
        status_code:
          type: integer
      required:
        - error
        - status_code
  responses:
    BadRequest:
      description: Bad request - invalid input
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
          example:
            error: "Bad Request"
            detail: "Invalid request format"
            status_code: 400
    NotFound:
      description: Resource not found
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
          example:
            error: "Not Found"
            detail: "Resource with specified ID does not exist"
            status_code: 404
    ValidationError:
      description: Validation error - invalid field values
      content:
        application/json:
          schema:
            type: object
            properties:
              error:
                type: string
              validation_errors:
                type: array
                items:
                  type: object
                  properties:
                    field:
                      type: string
                    message:
                      type: string
          example:
            error: "Validation Error"
            validation_errors:
              - field: "title"
                message: "Title must be at least 1 character"